+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # Change Log
+
+
+## Unreleased
+
+### Added
+
+- [go-kit](https://github.com/go-kit/kit/tree/master/log) logger constructor